pyproject.toml
[project]
name = "fastapi-football"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.121.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.9.0.post0",
    "sqlmodel>=0.0.27",
]
